fix(Ingredient): `name` only blank chars
Preciously `.isEmpty()` was called on the `name` parameter, but since it
returns true only if `.length` is zero.
Change to `.isBlank()` as it detects if the `String` is completely empty
or only contain white space and such case would be invalid for the
`name` field.

NOTE: Not all tests pass.