feat(datastorage): deserialize - one step closer
Test throw `ClassCastException` with below traceback:
java.lang.ClassCastException: class domain.Recipe cannot be cast to class java.util.Collection

The problem seems to be related to the `write` method and not the `read`
method. Further investigation will be done, and left a "TODO:" comment
to indicate place for possible bug.

NOTE: Not passing tests - see "TODO:" comment for leads