fix(datastorage): serialize and deserialize
Serialize entire `Collection` and write to file instead of writing a
single object to the file.
The logical error occurred due to the `read` method tried to type cast
the deserialized input stream to a `Collection`, which could maybe be
solved syntactically by `Recipe` implementing the `Collection`
interface - but this seems illogical.

NOTE: Passing all tests